requirements.txt
ddt==1.2.1
Jinja2==2.9.5
MarkupSafe==1.1.1
nose==1.3.7
PyYAML==5.1.1
selenium==3.141.0
urllib3==1.25.3
xlrd==1.2.0
